Much smoother movement except for changing direction
Basically just swapped OnChange events for polling. To be fair it gets hit a LOT so polling isn't bad. I was getting the impression that due to the curved smoothing on MovementAxis controls it was basically hitting the event insanely quickly from all the tiny step vector changes. That's not beneficial, it's slow, and often the actual tick of movement only triggers on a tiny number of those changed values anyway.